requirements.txt
fastapi
uvicorn
requests
python-multipart
chromadb
python-dotenv
prometheus_client
llama-cpp-python
wikipedia
onnxruntime
transformers
